requirements.txt
chardet==5.2.0
et_xmlfile==2.0.0
imbox==0.9.8
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
pyzbar==0.1.9
reportlab==4.3.0
six==1.17.0
tk==0.1.0
tzdata==2025.1
